+++ docs/changelog/changelog.rst
@@ -2,9 +2,13 @@
 
 develop
 -----------------
+* Better error message when config substitution variable is missing
+* removed an unused directory in the GE folder
+* removed obsolete config error handling
 * Docs typo fixes
 * Jupyter notebook improvements
 * `great_expectations init` improvements
+
 
 0.9.3
 -----------------